refactor(compiler): refactor tests
