+++ src/compiler/compiler_tests.rs
@@ -2,71 +2,14 @@
 #[cfg(test)]
 pub mod tests {
 
-    use std::rc::Rc;
-
     use crate::{
         compiler::{
-            code::{Instructions, Opcode},
-            test_utils::check_instructions,
-            Compiler,
+            code::Opcode,
+            test_utils::{flatten_instructions, run_compiler, CompilerTestCase},
         },
-        object::{
-            test_utils::check_constants,
-            {CompiledFunction, Object},
-        },
-        parser::parse,
+        object::Object,
     };
 
-    struct CompilerTestCase {
-        input: String,
-        expected_constants: Vec<Object>,
-        expected_instructions: Instructions,
-    }
-
-    fn flatten_instructions(instructions: Vec<Instructions>) -> Instructions {
-        let mut res = Instructions::default();
-        for instruction in instructions {
-            res.append(instruction);
-        }
-        res
-    }
-
-    fn flatten_u8_instructions(instructions: Vec<Instructions>) -> Vec<u8> {
-        let mut res = vec![];
-        for instruction in instructions {
-            res.append(&mut instruction.data.clone());
-        }
-        res
-    }
-
-    fn run_compiler(tests: Vec<CompilerTestCase>) {
-        for test in tests {
-            println!("Testing input: {}", test.input);
-            let program = parse(&test.input);
-
-            let mut compiler = Compiler::new();
-
-            match compiler.compile(program) {
-                Ok(_) => {
-                    let bytecode = compiler.bytecode();
-                    println!(
-                        "want {}, got {}",
-                        test.expected_instructions, bytecode.instructions
-                    );
-                    check_instructions(&bytecode.instructions, &test.expected_instructions);
-                    check_constants(
-                        &bytecode.constants,
-                        &test
-                            .expected_constants
-                            .iter()
-                            .map(|x| Rc::new(x.clone()))
-                            .collect(),
-                    );
-                }
-                Err(err) => panic!("compiler error: {err}"),
-            }
-        }
-    }
     #[test]
     fn test_integer_arithemtic() {
         let tests = vec![
@@ -572,602 +515,6 @@ pub mod tests {
     }
 
     #[test]
-    fn test_funtions() {
-        let tests = vec![
-            CompilerTestCase {
-                input: "fn() { return 5 + 10; }".to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(5),
-                    Object::INTEGER(10),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::Constant.make(vec![1]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 0,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![2, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "fn() { 5 + 10; }".to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(5),
-                    Object::INTEGER(10),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::Constant.make(vec![1]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 0,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![2, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "fn() { 1; 2 }".to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(1),
-                    Object::INTEGER(2),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::Pop.make(vec![]),
-                            Opcode::Constant.make(vec![1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 0,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![2, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "fn() { 24 }()".to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(24),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 0,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![1, 0]),
-                    Opcode::Call.make(vec![0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "let noArg = fn() { 24 }; noArg();".to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(24),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 0,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![1, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Call.make(vec![0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "let oneArg = fn(a) {}; oneArg(24);".to_string(),
-                expected_constants: vec![
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![Opcode::Return.make(vec![0])]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::INTEGER(24),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![0, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Constant.make(vec![1]),
-                    Opcode::Call.make(vec![1]),
-                    Opcode::Pop.make(vec![0]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "let manyArg = fn(a, b, c) { }; manyArg(24, 25, 26);".to_string(),
-                expected_constants: vec![
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![Opcode::Return.make(vec![0])]),
-                        num_locals: 3,
-                        num_parameters: 3,
-                    }),
-                    Object::INTEGER(24),
-                    Object::INTEGER(25),
-                    Object::INTEGER(26),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![0, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Constant.make(vec![1]),
-                    Opcode::Constant.make(vec![2]),
-                    Opcode::Constant.make(vec![3]),
-                    Opcode::Call.make(vec![3]),
-                    Opcode::Pop.make(vec![0]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "let oneArg = fn(a) { a; }; oneArg(24);".to_string(),
-                expected_constants: vec![
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::ReturnValue.make(vec![0]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::INTEGER(24),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![0, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Constant.make(vec![1]),
-                    Opcode::Call.make(vec![1]),
-                    Opcode::Pop.make(vec![0]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "let manyArg = fn(a, b, c) { a; b; c; }; manyArg(24, 25, 26);".to_string(),
-                expected_constants: vec![
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Pop.make(vec![0]),
-                            Opcode::GetLocal.make(vec![1]),
-                            Opcode::Pop.make(vec![0]),
-                            Opcode::GetLocal.make(vec![2]),
-                            Opcode::ReturnValue.make(vec![0]),
-                        ]),
-                        num_locals: 3,
-                        num_parameters: 3,
-                    }),
-                    Object::INTEGER(24),
-                    Object::INTEGER(25),
-                    Object::INTEGER(26),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![0, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Constant.make(vec![1]),
-                    Opcode::Constant.make(vec![2]),
-                    Opcode::Constant.make(vec![3]),
-                    Opcode::Call.make(vec![3]),
-                    Opcode::Pop.make(vec![0]),
-                ]),
-            },
-        ];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_function_with_no_return_value() {
-        let tests = vec![CompilerTestCase {
-            input: "fn() { }".to_string(),
-            expected_constants: vec![Object::COMPILEDFUNCTION(CompiledFunction {
-                instructions: flatten_u8_instructions(vec![Opcode::Return.make(vec![])]),
-                num_locals: 0,
-                num_parameters: 0,
-            })],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::Closure.make(vec![0, 0]),
-                Opcode::Pop.make(vec![]),
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_let_statements_scope() {
-        let tests = vec![
-            CompilerTestCase {
-                input: r#"
-                let num = 55;
-                fn() { num }"#
-                    .to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(55),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetGlobal.make(vec![0]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 0,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Constant.make(vec![0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::Closure.make(vec![1, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: r#"
-                fn() { 
-                    let num = 55;
-                    num
-                }"#
-                .to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(55),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::SetLocal.make(vec![0]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![1, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: r#"
-                fn() { 
-                    let a = 55;
-                    let b = 77;
-                    a + b
-                }"#
-                .to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(55),
-                    Object::INTEGER(77),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::SetLocal.make(vec![0]),
-                            Opcode::Constant.make(vec![1]),
-                            Opcode::SetLocal.make(vec![1]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::GetLocal.make(vec![1]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 2,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![2, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-        ];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_builtins() {
-        let tests = vec![
-            CompilerTestCase {
-                input: "len([]); push([], 1);".to_string(),
-                expected_constants: vec![Object::INTEGER(1)],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::GetBuiltin.make(vec![0]),
-                    Opcode::Array.make(vec![0]),
-                    Opcode::Call.make(vec![1]),
-                    Opcode::Pop.make(vec![]),
-                    Opcode::GetBuiltin.make(vec![4]),
-                    Opcode::Array.make(vec![0]),
-                    Opcode::Constant.make(vec![0]),
-                    Opcode::Call.make(vec![2]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: "fn() { len([]); }".to_string(),
-                expected_constants: vec![Object::COMPILEDFUNCTION(CompiledFunction {
-                    instructions: flatten_u8_instructions(vec![
-                        Opcode::GetBuiltin.make(vec![0]),
-                        Opcode::Array.make(vec![0]),
-                        Opcode::Call.make(vec![1]),
-                        Opcode::ReturnValue.make(vec![]),
-                    ]),
-                    num_locals: 0,
-                    num_parameters: 0,
-                })],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![0, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-        ];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-
-    fn test_closures() {
-        let tests = vec![
-            CompilerTestCase {
-                input: r#"
-                fn(a){
-                    fn(b){
-                        a + b
-                    }
-                }"#
-                .to_string(),
-                expected_constants: vec![
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetFree.make(vec![0]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Closure.make(vec![0, 1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![1, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: r#"
-                    fn(a) {
-                        fn(b) {
-                            fn(c) {
-                                a + b + c
-                            }
-                        }
-                    };"#
-                .to_string(),
-
-                expected_constants: vec![
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetFree.make(vec![0]),
-                            Opcode::GetFree.make(vec![1]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetFree.make(vec![0]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Closure.make(vec![0, 2]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Closure.make(vec![1, 1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![2, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: r#"
-                    let global = 55;
-                   fn() {
-                        let a = 66;
-                        fn() {
-                            let b = 77;
-                            fn() {
-                                let c = 88;
-                                global + a + b + c;
-                            }
-                        }
-                    }
-                    "#
-                .to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(55),
-                    Object::INTEGER(66),
-                    Object::INTEGER(77),
-                    Object::INTEGER(88),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![3]),
-                            Opcode::SetLocal.make(vec![0]),
-                            Opcode::GetGlobal.make(vec![0]),
-                            Opcode::GetFree.make(vec![0]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::GetFree.make(vec![1]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Add.make(vec![]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 0,
-                    }),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![2]),
-                            Opcode::SetLocal.make(vec![0]),
-                            Opcode::GetFree.make(vec![0]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Closure.make(vec![4, 2]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 0,
-                    }),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Constant.make(vec![1]),
-                            Opcode::SetLocal.make(vec![0]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Closure.make(vec![5, 1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Constant.make(vec![0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::Closure.make(vec![6, 0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-        ];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_recursive_functions() {
-        let tests = vec![
-            CompilerTestCase {
-                input: r#"
-                let countDown = fn(x){
-                    countDown(x - 1);
-                };
-                countDown(1);"#
-                    .to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(1),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::CurrentClosure.make(vec![]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::Sub.make(vec![]),
-                            Opcode::Call.make(vec![1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::INTEGER(1),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![1, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Constant.make(vec![2]),
-                    Opcode::Call.make(vec![1]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-            CompilerTestCase {
-                input: r#"
-                let wrapper = fn() {
-                    let countDown = fn(x) {
-                        countDown(x - 1);
-                    };
-                    countDown(1);
-                };
-                wrapper();
-                "#
-                .to_string(),
-                expected_constants: vec![
-                    Object::INTEGER(1),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::CurrentClosure.make(vec![]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Constant.make(vec![0]),
-                            Opcode::Sub.make(vec![]),
-                            Opcode::Call.make(vec![1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 1,
-                    }),
-                    Object::INTEGER(1),
-                    Object::COMPILEDFUNCTION(CompiledFunction {
-                        instructions: flatten_u8_instructions(vec![
-                            Opcode::Closure.make(vec![1, 0]),
-                            Opcode::SetLocal.make(vec![0]),
-                            Opcode::GetLocal.make(vec![0]),
-                            Opcode::Constant.make(vec![2]),
-                            Opcode::Call.make(vec![1]),
-                            Opcode::ReturnValue.make(vec![]),
-                        ]),
-                        num_locals: 1,
-                        num_parameters: 0,
-                    }),
-                ],
-                expected_instructions: flatten_instructions(vec![
-                    Opcode::Closure.make(vec![3, 0]),
-                    Opcode::SetGlobal.make(vec![0]),
-                    Opcode::GetGlobal.make(vec![0]),
-                    Opcode::Call.make(vec![0]),
-                    Opcode::Pop.make(vec![]),
-                ]),
-            },
-        ];
-
-        run_compiler(tests);
-    }
-
-    #[test]
     fn test_shadowing_with_itself() {
         let tests = vec![CompilerTestCase {
             input: r#"
@@ -1183,161 +530,6 @@ pub mod tests {
                 Opcode::Constant.make(vec![1]),
                 Opcode::Add.make(vec![]),
                 Opcode::SetGlobal.make(vec![0]),
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_while_statements() {
-        let tests = vec![CompilerTestCase {
-            input: r#"
-                    while (true){
-                        puts("yes");
-                    }
-                    "#
-            .to_string(),
-            expected_constants: vec![Object::STRING("yes".to_string())],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::True.make(vec![]),            // 000
-                Opcode::JumpNotTruthy.make(vec![15]), // 001
-                Opcode::GetBuiltin.make(vec![5]),     // 004
-                Opcode::Constant.make(vec![0]),       // 006
-                Opcode::Call.make(vec![1]),           // 009
-                Opcode::Pop.make(vec![]),             // 011
-                Opcode::Jump.make(vec![0]),           // 012
-                                                      // 015
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_break_in_while() {
-        let tests = vec![CompilerTestCase {
-            input: r#"
-                    while (true){
-                        break;
-                    }
-                    "#
-            .to_string(),
-            expected_constants: vec![],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::True.make(vec![]),            // 000
-                Opcode::JumpNotTruthy.make(vec![10]), // 001
-                Opcode::Jump.make(vec![10]),          // 004
-                Opcode::Jump.make(vec![0]),           // 007
-                                                      // 010
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_nested_breaks_in_while() {
-        let tests = vec![CompilerTestCase {
-            input: r#"
-                    while (true){
-                        while (true){
-                            break;
-                        }
-                        break;
-                    }
-                    "#
-            .to_string(),
-            expected_constants: vec![],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::True.make(vec![]),            // 000
-                Opcode::JumpNotTruthy.make(vec![20]), // 001
-                Opcode::True.make(vec![]),            // 004
-                Opcode::JumpNotTruthy.make(vec![14]), // 005
-                Opcode::Jump.make(vec![14]),          // 008
-                Opcode::Jump.make(vec![4]),           // 011
-                Opcode::Jump.make(vec![20]),          // 014
-                Opcode::Jump.make(vec![0]),           // 017
-                                                      // 020
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-    #[test]
-    fn test_continue_in_while() {
-        let tests = vec![CompilerTestCase {
-            input: r#"
-                        while (true){
-                            continue;
-                        }
-                        "#
-            .to_string(),
-            expected_constants: vec![],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::True.make(vec![]),            // 000
-                Opcode::JumpNotTruthy.make(vec![10]), // 001
-                Opcode::Jump.make(vec![0]),           // 004
-                Opcode::Jump.make(vec![0]),           // 007
-                                                      // 010
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_nested_continue_in_while() {
-        let tests = vec![CompilerTestCase {
-            input: r#"
-                    while (true){
-                        while (true){
-                            continue;
-                        }
-                        continue;
-                    }
-                    "#
-            .to_string(),
-            expected_constants: vec![],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::True.make(vec![]),            // 000
-                Opcode::JumpNotTruthy.make(vec![20]), // 001
-                Opcode::True.make(vec![]),            // 004
-                Opcode::JumpNotTruthy.make(vec![14]), // 005
-                Opcode::Jump.make(vec![4]),           // 008
-                Opcode::Jump.make(vec![4]),           // 011
-                Opcode::Jump.make(vec![0]),           // 014
-                Opcode::Jump.make(vec![0]),           // 017
-                                                      // 020
-            ]),
-        }];
-
-        run_compiler(tests);
-    }
-
-    #[test]
-    fn test_continue_and_break_in_while() {
-        let tests = vec![CompilerTestCase {
-            input: r#"
-                    while (true){
-                        while (true){
-                            continue;
-                        }
-                        break;
-                    }
-                    "#
-            .to_string(),
-            expected_constants: vec![],
-            expected_instructions: flatten_instructions(vec![
-                Opcode::True.make(vec![]),            // 000
-                Opcode::JumpNotTruthy.make(vec![20]), // 001
-                Opcode::True.make(vec![]),            // 004
-                Opcode::JumpNotTruthy.make(vec![14]), // 005
-                Opcode::Jump.make(vec![4]),           // 008
-                Opcode::Jump.make(vec![4]),           // 011
-                Opcode::Jump.make(vec![20]),          // 014
-                Opcode::Jump.make(vec![0]),           // 017
-                                                      // 020
             ]),
         }];
